Remove profile memory from LoCoMo benchmark using agent
