+++ evaluation/locomo/episodic_agent/README.md
@@ -1,6 +1,6 @@
 ## Add conversations to memory
 ```sh
-python run_experiments.py --method add --dataset path/to/locomo10.json
+python locomo_ingest.py --data-path path/to/locomo10.json
 ```
 
 ## Search memory and answer questions
@@ -21,12 +21,5 @@ python generate_scores.py --input_path evaluation.json
 
 ## Delete data
 ```sh
-memmachine-sync-profile-schema --delete \
-    -- host <HOST> \ # or use environment variable POSTGRES_HOST
-    -- port <PORT> \ # or use environment variable POSTGRES_PORT
-    -- user <USER> \ # or use environment variable POSTGRES_USER
-    -- password <PASSWORD> \ # or use environment variable POSTGRES_PASSWORD
-    -- database <DATABASE> # or use environment variable POSTGRES_DB
-
 python locomo_delete.py --data-path path/to/locomo10.json
 ```
