+++ evaluation/locomo/episodic_agent/run_experiments.py
@@ -29,13 +29,10 @@ def main():
         f"\nBase Url: {args.base_url}"
     )
 
-    from memmachine_locomo import MemMachineAdd, MemMachineSearch
+    from memmachine_locomo import MemMachineSearch
 
     dataset = args.dataset
-    if args.method == "add":
-        im = MemMachineAdd(dataset)
-        asyncio.run(im.process_all_conversations(base_url=args.base_url))
-    elif args.method == "search":
+    if args.method == "search":
         im = MemMachineSearch()
         asyncio.run(im.process_data_file(dataset, base_url=args.base_url))
     else:
